+++ src/main/java/nextstep/auth/ui/AuthController.java
@@ -1,0 +1,23 @@
+package nextstep.auth.ui;
+
+import nextstep.auth.application.AuthService;
+import nextstep.auth.application.dto.TokenRequest;
+import nextstep.auth.application.dto.TokenResponse;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class AuthController {
+
+    private final AuthService authService;
+
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login/token")
+    public TokenResponse login(@RequestBody final TokenRequest request) {
+        return authService.login(request);
+    }
+}
