+++ src/main/java/nextstep/member/ui/MemberController.java
@@ -1,6 +1,6 @@
 package nextstep.member.ui;
 
-import nextstep.member.application.AuthService;
+import nextstep.auth.application.AuthService;
 import nextstep.member.application.MemberService;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
